requirements.txt
pytest==8.4.2
selenium==4.35.0
allure-pytest==2.15.3
allure-python-commons==2.15.0